[needs binaries] combine Clang, GCC, Binutils, and XCodeCLITools to form the NativeToolchain subsystem and consume it in BuildLocalPythonDistributions for native code (#5490)
### Problem

`python_dist()` allows the use of native extension modules, but it currently uses whatever compiler, linker, libraries, *etc* that distutils can find first on the host to compile the native code. Providing these disparate elements of the native code compilation toolchain in Pants means users don't have to install it themselves, and Pants can ensure that the provided compiler supports whatever features we want to provide to developers of native extension modules with `python_dist()`.

### Solution

This PR uses the LLVM subsystem defined in #5471 (now called `Clang`) along with multiple other subsystems to compile native code in a setup.py-based project declared through a `python_dist()` target.

### Result

We provide `clang` and `gcc` for all platforms, `binutils` on Linux to link, and we wrap the XCode command line developer tools to link on OSX. When a `python_dist()` target is built through the `BuildLocalPythonDistributions` task, it pulls in the bootstrapped toolchain and prepends it to the `PATH` so that our provided compiler is preferred (in addition to scrubbing `CC` and `CXX`). True sandboxed execution is preferable, but for now this at least guarantees that the provided compiler is used, which satisfies the goals stated at the top.