@rules as coroutines (#5580)
### Problem

The "complex" `Selector` types used to select `@rule` inputs for subjects other than the current subject of an `@rule` have proven to be entirely too complicated. Additionally, their lack of flexibility meant that we needed to introduce many intermediate types to provide links between `@rules` in cases where imperative logic was needed for filtering or conditionals.

The only remaining motivating factor for these complex selectors was that they allow for closed-world compilation of the `@rule` graph, which avoids the need for error checking in rule bodies and provides for easier composition of sets of `@rules`.

### Solution

After experimenting with using coroutines (ie: functions containing `yield` statements that receive a result on their left hand side), it became clear both that:

1. Coroutines were universally easier to understand than the Selector DSL
2. The dependency requests made by coroutines could still easily be statically checked via a small amount of runtime introspection.

And so, this change removes all usage of `SelectProjection` and most usage of `SelectDependencies` in favor of `@rule` coroutines. It does not yet remove the implementation of those `Selectors` (although doing so will shed a massive amount of complexity from the codebase).

### Result

`@rules` are still statically checked, but are now much easier to write. Performance is unaffected.