brfs: FUSE filesystem exposing the Store and remote CAS
Exposes:
 * ${mount_point}/digest - contains a file named sha256-length for each
   file digest. Does not show anything in response to `readdir`, but
   opening any particular file (of a known digest) will work.
 * ${mount_point}/directory - contains a directory named sha256-length
   for each directory digest. Does not show anything in response to
   `readdir`, but `readdir` on any particular directory (of a known
   digest) will work.

Caveats:
 * Not at all performance optimised. In particular, all filesystem
   operations happen single-threaded, including performing expensive I/O
   operations like fetching from a remote CAS or reading from LMDB.
   The underlying implementation is Futures-based, and the expensive
   operations are easily parallelisable, it just hasn't been done yet :)

This uses the low-level https://github.com/zargony/rust-fuse which
requires manual inode management.

There is also a higher-level https://github.com/wfraser/fuse-mt which I
discovered after putting this together, but which I'm consciously not
using. Its benefits would be:
 1. It operates on paths rather than inodes. This would be pretty
    convenient, as the inode cache in this implementation is a little
    annoying.
 2. It runs some operations in parallel by not having every function
    take a &mut self. I haven't looked at how parallel it ends up being.
 3. Its readdir API avoids needing to think about pagination. I ended up
    coincidentally re-inventing their API here :)

But its big drawback is that functions are required to return values
(e.g. from read) rather than call callbacks, which means that it's hard
for us to take advantage of our Futurey implementation in Store, as
we'd need to block in every call.

Both projects have talked about wanting to support Futures as a more
first-class concept at some point in a nebulous future, but neither has
a concrete plan for doing so.