type-check specific datatype fields concisely and remove the class name argument (#5723)
### Problem

See #5716. We had to introduce a ton of boilerplate to check argument types in some very simple usage of `datatype` subclasses in tests in #5703. This is one way to make that easier. 

### Solution

- Move `TypeConstraint` and subclasses from `src/python/pants/engine/addressable.py` to `src/python/pants/util/objects.py`, where they probably should have been anyway, to avoid an import cycle.
- Remove the class name argument from `datatype()` (and update all call sites).
- Allow specifying a tuple `('field_name', FieldType)` in the list of fields to `datatype()` to have the field type-checked by an `Exactly(FieldType)` type constraint at construction.
- Override `__str__()` and `__repr__()` for `datatype` objects, and add testing for the str and repr.
- Add testing for informative error messages in `datatype` construction.

### Result

We can now concisely declare `datatype` objects without having to repeat the class name, and can opt-in to a type check for specific fields.