Introduce Resettable (#5770)
Because we fork without execing a lot, we often need to reset things
that have references to background threads. This provides a handy
wrapper for doing so.

If this looks good, I will also apply it to the
process_execution::remote::CommandRunner.

I can also apply it to the ResettablePool which we already have, for
re-use (it would add an additional Arc clone per CpuPool operation, but
this is probably fine because the CpuPool operations are already by
definition somewhat heavyweight).