[pytest runner] re-add --options flag as a shlexed list of strings (#5790)
### Problem

`--options` was removed from pytest when cleaning up some issues with passthrough args (https://github.com/pantsbuild/pants/pull/5594/). This made it more difficult to pass arbitrary args for pytest runs for automated use cases or to have options set in `pants.ini`. It makes it more difficult because now those options must be collated and appended to the end of the command in all cases. If the command construction code splits up how different flags are populated, it can be difficult to ensure that a particular flag's contents always ends up at the end.

### Solution

Bring back the options flag, but only shlex it and not the passthrough args.