Add python_app support (#5704)
Pants supports the "app" concept - a binary + other loose files - but currently only has the `jvm_app` implementation. Here we add `python_app` which is given a `python_binary` and `bundles` to include in an archive. This enables users to easily create a pex along with other files needed in production and create a deployable archive.