Add a facility for exposing metrics from the native engine (#5808)
### Problem

As described in #5736, some of the metrics we've recently added introduce dependencies on having constructed a `BuildGraph`. In order to unblock removing the `BuildGraph` requirement in #4769, we'd like to replace those metrics with forward compatible metrics computed in the native engine.

### Solution

Split out a `Session` type which represents a scope for metrics capture. Currently the `Session` records the root requests that were made during its lifetime, which allows it to later determine which nodes are reachable from those roots. In future, additional metrics could be captured in the `Session` object at the beginning (ie, when it is created) or "end" (when `session.metrics()` is called) of the `Session`.

The python objects representing the scheduler were renamed to `Scheduler` and `SchedulerSession` to represent this split. `pantsd` reuses a `Scheduler` for its entire lifetime, but creates a new `SchedulerSession` per client request, which is then used post fork as well.

### Result

The `affected_target_file_count` metric was replaced with `affected_file_count` metric, based on the count of digests reachable from all of the roots in a session. Fixes #5736 and unblocks adding further scheduler metrics.