Use write_all not write
write may not write all bytes, and will not return an error if only a
partial write was performed.

This was surfaced by clippy; I don't think clippy is stable enough to
turn on on CI, but I may start running occasional clippy checks to look
for issues / things to clean up.