Improve logging/handling of signaled, killed, and terminated tests (#5859)
### Problem

While debugging #5611, I determined that when a test was killed by a signal (and thus had a negative return value from `poll()`), `TestRunnerTaskMixin` incorrectly interpreted the failure as a still-living and hung process, and would later try to kill it. There were a few broken tests allowing for this, but primarily: `test_pytest_run_timeout_cant_terminate` was not waiting long enough for the tested-test to start up, and so was killing it before it had its signal handler in place.

Additionally, the usage of a `Timer` + `poll()` to implement test termination meant that we were guaranteed to wait the full `timeout_terminate_wait` time (10 seconds by default) before `poll()`ing to see whether the test had exited.

### Solution

1. Interpret `poll() == None` as a still running process, and `poll() < 0` as a process killed by a signal.
2. Report processes killed by signals before our initial attempt to kill them (which should expose a SIGSEV in #5611).
3. Rather than a timer, use `wait()` between `terminate()` and `kill()`, which avoids unnecessary sleeping.

### Result

Improved debuggability for tests that exit abnormally.