[rust/engine] add timeout \ description for grpc process requests (#5632)
Fixes #5504

On responses from the server, instead of always continuing, we'll now timeout if the incomplete responses take more than the timeout. The default timeout in the patch is 15 minutes.

It does this by introducing a timeout to ExecuteProcessRequest that counts from when the grpc request goes out and fails with an error result if the timeout is exceeded while checking the status of the response.

It also adds a description field and includes it in error messages so that errors can have a human readable description of the request.

It does not add a timeout implementation to local executions, #5833 is the issue for that.