Sort list goal output (#5872)
Currently the output of `./pants list` is not sorted, which is confusing for users. Collecting the results and sorting them before displaying the output does not appear to have a noticable impact on performance, which takes ~4 seconds in the pants repo at present, and its basically the same with the sort, since the bulk of the time is building the graph.