Fix visualize-to option post addition of Sessions (#5885)
### Problem

When `Sessions` were added, the "static" reference to the most recent `ExecutionRequest` was removed from the `Scheduler`. But that reference was still in use by `_maybe_visualize` (used in the background when `--native-engine-visualize-to` was set)... which didn't have an integration test.

### Solution

Since visualizing "the most recent request" didn't really make sense in all contexts where `_maybe_visualize` was used anyway, switch to visualizing all roots that have been requested within a `Session`. Cover this with an integration test.

### Result

`--native-engine-visualize-to`  works, and is protected against future breakage.