Robustify go_protobuf_library (#5887)
Now that `go_protobuf_library` has been added, I found a couple minor issues in real-world use-cases. Not all protobuf IDL files contain a Go namespace, so we should fall back to the package. Also, we should use the Go namespace as-is in the path names as sometimes there are path elements that contain a `.` (e.g.: `github.com`) which is a Go thing.