Put the ChangeCalculator implementation next to TargetRootsCalculator (#5917)
### Problem

While working on #5912, I found the code related to target calculation (currently just used along the `--changed` option and thus `ChangeCalculator` very disperse.

### Solution

After discussing it with @stuhood, we came to the conclusion that bringing the code together and simplifying the dependencies would make this part of the codebase easier to work with. And thus this PR aims to do just that.

### Result

The user should not be aware of this change. It is internal codebase refactoring.

PS. I skipped the pre-commit hook due to a bug with my isort.sh script, so it may fail on CI before I fix the import order.