Enable fromfile support for --owner-of and increase test coverage (#5948)
### Problem

The new `--owner-of` option was broken in the context of `pantsd`, but didn't have test coverage due to the `--changed` and `--owner-of` tests not running under `pantsd`. Additionally, `fromfile` support is useful for this option, but was not enabled. 

### Solution

Mark some integration tests as needing to run under the daemon, and enable `fromfile` support for `--owner-of`.