Re-shade zinc to avoid classpath collisions with annotation processors. (#5953)
### Problem

Zinc used to be shaded before the `1.x.y` upgrade (#4729), but shading was removed due to an overabundance of optimism. When testing the zinc upgrade internally, we experienced a classpath collision between an annotation processor and zinc (in guava, although zinc has many other dependencies that could cause issues).

### Solution

Shade zinc, and ensure that our annotation processor uses a very old guava in order to attempt to force collisions in future.