Record start times per graph node and expose a method to summarize them. (#5964)
### Problem

In order to display "significant" work while it is running on the engine, we need to compute interesting, long-running leaves.

### Solution

Record start times per entry in the `Graph`, and add a method to compute a graph-aware top-k longest running leaves. We traverse the graph "longest running first", and emit the first `k` leaves we encounter.

### Result

While this will almost certainly need further edits to maximize it's usefulness, visualization can begun to be built atop of it.