Enable passing option sets to the compiler (#6065)
### Problem

The issue at #5360 

### Solution

I've added another option, `--compiler-option-sets`, which takes a list of option sets which then get translated to the actual flags that you want to pass to the compiler. These flags should be defined in `pants.ini` or through the command line.

### Result

Users can now enable compiler flags on a project or session basis instead of the whole repo.

PS. This is a work in progress, existing tests are failing and new ones have not been implemented. Making the PR now so that it can be reviewed as it progresses.