Add libc search noop option (#6122)
### Problem

Our `LibcDev` subsystem searches for `crti.o`, a file necessary to create executables on Linux. We use this in order to test both clang and gcc in `test_native_toolchain.py`. This is not needed except within Pants CI right now, and if the host system doesn't contain this file, it will always error out.

### Solution

- Add `--enable-libc-search` to `NativeToolchain`, defaulting to `False`, so that systems without this currently-unnecessary file don't need to worry about this implementation detail.
- Create `@platform_specific` decorator for individual tests which should only run on a specific platform to validate the expected behavior of this flag.

### Result

Pants invocations using the native toolchain on Linux hosts without a `crti.o` do not error out.