Fix issue of os.environ expecting bytes vs unicode in Py2 vs Py3 (#6222)
### Problem
In Py2, `os.environ` expects either bytes or ASCII strings. So, to pass it unicode, you must encode into bytes and pass the bytes, which is why we have several helper functions.

In Py3, it expects unicode and will crash upon receiving bytes.

We didn't catch this because there is no backport for the `os` module, and this is the first time running `contextutil` with Py3.

### Solution
If Py2, keep existing conversion from unicode to bytes. If Py3, don't modify the passed in unicode.

### Testing
CI is only going to prove this doesn't break Py2. It has no guarantees Py3 will work.. I got the unit tests to work, but there may still be other issues I don't know how to detect because we can't yet run CI in Py3 (too many other issues that have to be resolved)