Python 3 fixes - fix scm bytes vs unicode issues (#6257)
This folder had a lot of issues with intermixing unicode and bytes. 

While I think some of it can be refactored to use unicode (simpler to deal with), I tried to stop myself from doing this and to stay as true to the original semantics as possible.

Tests now pass locally in Py2 and Py3.