When python target compatibility is not set, use interpreter constraints. (#6284)
### Problem

Currently, `interpreter_constraints` are not used as the "default" value of `compatibility` for a python target, which means that:
```
./pants --python-setup-interpreter-constraints='["CPython>=3.5"]' test tests/python/pants_test/java:nailgun_io
```

...fails in an odd half-configured fashion (where the interpreter is bounded in some way, but not the dep resolution).

### Solution

Use `interpreter_constraints` as the default value of `compatibility` for a target. This aligns with #6250, and that patch should likely land as well.

### Result

```./pants --python-setup-interpreter-constraints='["CPython>=3.5"]' test tests/python/pants_test/java:nailgun_io```
...works, and uses python 3.

Fixes #5082.

----

Before this patch, in a repository with:
```
--python-setup-interpreter-constraints='["CPython>=2.7,CPython<3"]'
```
...individual targets were able to claim to be compatible with `CPython>3` while depending on targets with no declared compatibility, and things would just work.

After this patch, the default compatibility for a repository is specified via the constraints, and so rather than being completely unconstrained, a target with no compatibility argument would pick up the default. In the case above, that would cause a failure.

There are a few potential fixes for consumers:
1. don't pin interpreters in pants.ini, and only pin compatibility on binaries, or on targets where it is relevant
2. pin interpreters in pants.ini, but for a range wide enough to cover all of your binary and test target's compatibilities
3. pin interpreters for a "default" compatibility (ie, default to 2), and then explicitly clear the default on libraries to make them "agnostic": ie, `python_library(.., compatibility=[])` (or declare them to be wide enough to cover all consumers).